+++ interface/src/avatar/AvatarMotionState.h
@@ -62,7 +62,7 @@ public:
 
     void addDirtyFlags(uint32_t flags) { _dirtyFlags |= flags; }
 
-    virtual void computeCollisionGroupAndMask(int16_t& group, int16_t& mask) const override;
+    virtual void computeCollisionGroupAndMask(int32_t& group, int32_t& mask) const override;
 
     friend class AvatarManager;
     friend class Avatar;
